Dynamic protocol listener refactor
Introduce a new functio that allow easilly add/remove
listener, which can be used across the whole protocol
stack. One only need to create the listener and the
coressponding helper which help to initialize the listener,
finally annotate the helper on target listener.

The is dynamic and extensiable without the needs to change
original protocol stack.