Fix the received messages executed in different thread are not synchronized
Since we need mantain serializability of message delivery and view delivery,
there is need to setup critica section for them. This is currently done via
a onProcessNo which get increase when service start, and decrease upon finish.
However when the service is executed in different thread (due to the need to
suspend thread, and this need to be done in different thread as GCM normally
do not tolerate it) then the noProcessNo is decrease even before the service
compelete. Solution would be using thread local and common executor to keep
them synchonized. Also fix a bug on NOTIFY_SUSPECT of FT protocol.