Recfactor of concurrent delivery service
The original notion suffering the issue when a serial
of messages m1, m2 and m3 arrive the sequencer with such
order, if m3 CDS m2 and m2 CDS m1 but m3 is not CDS m1,
then m3 may be concurrently deliveried with m1 as m3 and m2
would both marked as concurrent deliverable. This violate
casual consistency.

Amendment has been made to use a list of CDS, therefore each
message can examine if it can be concurrent delivery with all
the previous messages. The list would be flushed once the next
ordered sequence being assigned.